+++ test/GuardManager.t.sol
@@ -17,15 +17,15 @@ contract GuardManagerTest is Test {
         mockGuard = new MockGuard();
     }
 
-    function _createGuards() internal view returns(GuardDefinition[] memory) {
+    function _createGuards() internal view returns (GuardDefinition[] memory) {
         GuardParamType[] memory paramTypes = new GuardParamType[](1);
         paramTypes[0] = GuardParamType.Executor;
 
         GuardDefinition memory guard = GuardDefinition(
             address(mockGuard),
             "isWhitelisted(address)",
-            bytes32(uint(1)),
-            paramTypes, 
+            bytes32(uint256(1)),
+            paramTypes,
             new bytes32[](0),
             RequestType.Deposit,
             "=="
@@ -34,17 +34,17 @@ contract GuardManagerTest is Test {
         GuardDefinition memory guard2 = GuardDefinition(
             address(mockGuard),
             "isWhitelisted(address)",
-            bytes32(uint(1)),
-            paramTypes, 
+            bytes32(uint256(1)),
+            paramTypes,
             new bytes32[](0),
             RequestType.Withdrawal,
             "=="
         );
-         GuardDefinition memory guard3 = GuardDefinition(
+        GuardDefinition memory guard3 = GuardDefinition(
             address(mockGuard),
             "isWhitelisted(address)",
-            bytes32(uint(1)),
-            paramTypes, 
+            bytes32(uint256(1)),
+            paramTypes,
             new bytes32[](0),
             RequestType.Withdrawal,
             "=="
@@ -54,7 +54,7 @@ contract GuardManagerTest is Test {
         guards[0] = guard;
         guards[1] = guard2;
         guards[2] = guard3;
-        
+
         return guards;
     }
 
@@ -74,19 +74,14 @@ contract GuardManagerTest is Test {
 
     function testRunGuards() public {
         guardManager.setGuards(smartVaultId, _createGuards());
-        RequestContext memory context = RequestContext(
-            address(user),
-            address(user),
-            RequestType.Deposit,
-            new uint256[](0),
-            new address[](0)
-        );
+        RequestContext memory context =
+            RequestContext(address(user), address(user), RequestType.Deposit, new uint256[](0), new address[](0));
 
         vm.expectRevert("GuardManager::_checkResult: A-a, go back.");
         guardManager.runGuards(smartVaultId, context);
 
         mockGuard.setWhitelist(user, true);
-        
+
         guardManager.runGuards(smartVaultId, context);
     }
 }
